// all_projects/dev/health-backend-testing/testing-dashboard/backend/src/config/config.enum.ts
export enum sourceSystemTypesEnum {
  'KIPU' = 'kipu',
  'RITTEN' = 'ritten',
}

export enum EnvironmentList {
  clientDev = 'client_dev',
  production = 'production',
}

export enum DataSource {
  KIPU = 'KIPU',
  LOCAL = 'LOCAL',
  RITTEN = 'RITTEN',
}

export enum PatientDischargeStatus {
  DISCHARGED = 'DISCHARGED',
  IN_PATIENT = 'IN PATIENT',
}