// all_projects/dev/health-backend-testing/testing-dashboard/backend/src/lib/kipu/interface/kipu.interface.ts
export interface KipuDataInterface {
    casefile_id: string
    first_name: string
    middle_name: string
    last_name: string
    gender: string
    gender_identity: string
    dob: string
    race: string
    ethnicity: string
    address_street: string
    address_street2: string
    address_city: string
    state: string
    address_zip: string
    address_country: string
    email: string
    pre_admission_status: string
    admission_date: string
    discharge_date: string
    anticipated_discharge_date: any
    date_of_death: any
    cause_of_death: any
    discharge_or_transition_id: any
    discharge_or_transition_name: string
    discharge_type: string
    discharge_type_code: string
    first_contact_name: string
    referrer_name: string
    mr_number: string
    payment_method: string
    payment_method_category: string
    created_at: string
    last_updated_at: string
    diagnosis_codes: string
    pre_admission_status_id: any
    preferred_name: string
    alternate_phone: string
    sobriety_date: string
    ssn: string
    first_contact_date: string
    first_contact_information: string
    first_contact_relationship: string
    first_contact_rep_on_call: string
    patient_contacts: any[]
    phone: string
    level_of_care: string
    next_level_of_care: string
    next_level_of_care_date: any
    program: string
    bed_name: string
    room_name: string
    building_name: string
    location_id: number
    location_name: string
    external_apps: any[]
    ext_id_mappings: any[]
    record_source: string
    pronouns: string
    preferred_language: string
    alternate_addresses: any[]
    patient_tags: any[]
    preferred_contact: string
    maiden_name: string
    insurances01: string
}